made location selection user friendly
